WDD-18598 (SPARK-20640): Make RPC timeout and retry for shuffle registration configurable
Currently the shuffle service registration timeout and retry has been hardcoded. This works well for small workloads but under heavy workload when the shuffle service is busy transferring large amount of data we see significant delay in responding to the registration request, as a result we often see the executors fail to register with the shuffle service, eventually failing the job. We need to make these two parameters configurable.

* Updated `BlockManagerSuite` to test registration timeout and max attempts configuration actually works.